requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
blis==0.7.7
catalogue==2.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cymem==2.0.6
DAWG-Python==0.7.2
Django==4.0.5
django-environ==0.8.1
djangorestframework==3.13.1
docopt==0.6.2
idna==3.3
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.7
nltk==3.7
numpy==1.22.4
packaging==21.3
pandas==1.4.2
pathy==0.6.1
preshed==3.0.6
psycopg2==2.9.3
pydantic==1.8.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
regex==2022.6.2
requests==2.28.0
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
sklearn==0.0
smart-open==5.2.1
spacy==3.3.1
spacy-legacy==3.0.9
spacy-loggers==1.0.2
sqlparse==0.4.2
srsly==2.4.3
thinc==8.0.17
threadpoolctl==3.1.0
tqdm==4.64.0
typer==0.4.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
wasabi==0.9.1
